fix(error): change process when throwing error

+++ src/modules/http/index.ts
@@ -3,6 +3,7 @@ import { createWriteStream, copyFileSync, statSync } from 'fs'
 
 import { IDownloader, IOptions } from '@services/interfaces'
 import { removeFile, generateTempFilename } from '@libs/utils'
+import { ERROR } from '@libs/constants'
 
 import { getDestinationFromURL } from './utils'
 
@@ -62,36 +63,37 @@ class Main implements IDownloader {
     }
   }
 
-  async download(options: IOptions): Promise<void> {
+  download(options: IOptions): Promise<void> {
     return new Promise((resolve, rejects) => {
       // temporary destination until download finish
       this._dest = generateTempFilename()
-      try {
-        // request for downloading
-        fetch(options.url).then(res => {
-          const stream = createWriteStream(this._dest)
-          // read size from header
-          this._size = parseInt(res.headers.get('content-length'), 10)
-          // start write stream
-          res.body.pipe(stream)
-          stream.on('close', () => {
-            // this process is closed unproperly
-            if (!this._completed) {
-              removeFile(this._dest)
-            }
-          })
-          stream.on('finish', () => {
-            copyFileSync(this._dest, getDestinationFromURL(options.url, options.dest))
-            this._dest = options.dest
-            this._completed = true
-            resolve()
-          })
+      if (!options.url) {
+        throw new Error(ERROR.URL_IS_INVALID)
+      }
+      // request for downloading
+      fetch(options.url).then(res => {
+        const stream = createWriteStream(this._dest)
+        // read size from header
+        this._size = parseInt(res.headers.get('content-length'), 10)
+        // start write stream
+        res.body.pipe(stream)
+        stream.on('close', () => {
+          // this process is closed unproperly
+          if (!this._completed) {
+            removeFile(this._dest)
+          }
         })
-      } catch (e) {
-        // remove file is something wrong
+        stream.on('finish', () => {
+          copyFileSync(this._dest, getDestinationFromURL(options.url, options.dest))
+          this._dest = options.dest
+          this._completed = true
+          resolve()
+        })
+      }).catch(e => {
+        // remove file if something wrong happend
         removeFile(this._dest)
         rejects(e)
-      }
+      })
     })
   }
 }
